refactor(proxy): group auth_v2 into authn/authz/stages subpackages
Pure move (git mv, history preserved): the flat 17-file package hid the
architecture behind an alphabetical wall. Group by request phase so the module
is self-documenting:

  authn/   authenticators, jwt_claims, jwt_verifier, oauth2_introspection
  authz/   enforcer + model.conf, authorizer, route_map, policy_store, policy_admin
  stages/  enrichment, end_user, budgets
  (top)    entry, context, principal, management_endpoints, __init__

No content changes beyond rewired relative imports; tests moved to mirror the
layout. Behavior is identical — the same 129 tests pass, mypy/black/ruff clean,
and the package still imports with casbin/authlib absent.
